add deleteObject, fix return of getIndex

+++ app/src/main/cpp/WINAPI_KERNEL/SDK/include/Windows/Kernel/WindowsAPITable.h
@@ -29,7 +29,7 @@ typedef class Table {
             void skip(size_t amount);
             bool hasNext();
             Object * next();
-            int getIndex() const;
+            size_t getIndex() const;
         };
 
         Iterator getIterator();
